+++ RiskOfChaos/EffectHandling/EffectDispatchTimerType.cs
@@ -1,0 +1,8 @@
+﻿namespace RiskOfChaos.EffectHandling
+{
+    public enum EffectDispatchTimerType : sbyte
+    {
+        Unpaused,
+        Paused
+    }
+}